genovate/backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain
langchain-google-genai
google-cloud-aiplatform
google-cloud-bigquery
pydantic
pandas
pytest
